services/agents/recovery-coordinator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ranger-recovery-coordinator"
version = "0.1.0"
description = "Root orchestrator agent for the RANGER platform"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "RANGER Team", email = "ranger@example.com" }
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "google-adk>=0.1.0",
    "python-dotenv>=1.0.0",
    "ranger-agent-common",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
